requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-community
python-dotenv
pydantic